Add ui

+++ src/main/java/ru/BoshkaLab/repositories/AnswerRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import ru.BoshkaLab.entities.Answer;
 
 public interface AnswerRepository extends JpaRepository<Answer, Long> {
-    int countAllByEmployee_Id(long employeeId);
+    int countAllByEmployee_Id(int employeeId);
 }
